requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and acceleration
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0

# Dataset and data processing
datasets>=2.18.0
Pillow>=10.0.0
opencv-python>=4.8.0
decord>=0.6.0
av>=10.0.0

# Utilities
tqdm>=4.65.0
einops>=0.7.0
sentencepiece>=0.1.99

# Evaluation metrics
nltk>=3.8
rouge-score>=0.1.2
sacrebleu>=2.3.1
pandas>=2.0.0
tabulate>=0.9.0
requests>=2.31.0
gdown
# Optional: vLLM for faster inference
# vllm>=0.6.3

# Optional: pycocoevalcap for ScanQA
# git+https://github.com/salaniz/pycocoevalcap
